\name{mkRanefStructures}
\alias{mkRanefStructures}
\title{Make random effects structures}
\usage{
mkRanefStructures(grp, mm)
}
\arguments{
  \item{grp}{List of grouping factors (one for each term)}

  \item{mm}{List of model matrices (one for each term)}
}
\description{
Make random effects structures
}

