\name{mkTemplate}
\alias{mkTemplate}
\title{Make a single template for a relative covariance factor}
\usage{
mkTemplate(nc)
}
\arguments{
  \item{nc}{Number of columns in a dense model matrix for a
  particular random effects term}
}
\description{
Make a single template for a relative covariance factor
}
\examples{
mkTemplate(5)
}

