\name{Zsection}
\alias{Zsection}
\title{Create a section of a transposed random effects design matrix}
\usage{
  Zsection(grp, mm)
}
\arguments{
  \item{grp}{grouping factor}

  \item{mm}{dense model matrix}
}
\value{
  Dense section of a random effects design matrix
}
\description{
  Create a section of a transposed random effects design
  matrix
}

